upgrade ViewTotalGiftsModal
